pyproject.toml
[project]
name = "speculative-decoding"
version = "0.1.0"
description = "Speculative Decoding Research"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "huggingface-hub>=1.7.2",
    "python-dotenv>=1.2.2",
    "torch>=2.10.0",
    "transformers>=5.3.0",
]
